refactor: agrupar en carpetas las clases DTOs
