+++ src/main/java/asist/io/controller/CursoController.java
@@ -1,8 +1,8 @@
 package asist.io.controller;
 
-import asist.io.dto.CursoGetDTO;
-import asist.io.dto.CursoPatchDTO;
-import asist.io.dto.CursoPostDTO;
+import asist.io.dto.cursoDTO.CursoGetDTO;
+import asist.io.dto.cursoDTO.CursoPatchDTO;
+import asist.io.dto.cursoDTO.CursoPostDTO;
 import asist.io.exception.ModelException;
 import asist.io.service.ICursoService;
 import jakarta.validation.Valid;
