+++ src/main/java/asist/io/dto/cursoDTO/CursoGetDTO.java
@@ -1,4 +1,4 @@
-package asist.io.dto;
+package asist.io.dto.cursoDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,8 +11,10 @@ import java.io.Serializable;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class EstudianteGetDTO implements Serializable {
+public class CursoGetDTO implements Serializable {
     private String id;
-    private String lu;
     private String nombre;
+    private String descripcion;
+    private String carrera;
+    private String codigoAsistencia;
 }
