+++ src/main/java/asist/io/dto/cursoDTO/CursoPostDTO.java
@@ -1,4 +1,4 @@
-package asist.io.dto;
+package asist.io.dto.cursoDTO;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
