+++ src/main/java/asist/io/mapper/CursoMapper.java
@@ -1,8 +1,8 @@
 package asist.io.mapper;
 
-import asist.io.dto.CursoGetDTO;
-import asist.io.dto.CursoPatchDTO;
-import asist.io.dto.CursoPostDTO;
+import asist.io.dto.cursoDTO.CursoGetDTO;
+import asist.io.dto.cursoDTO.CursoPatchDTO;
+import asist.io.dto.cursoDTO.CursoPostDTO;
 import asist.io.entity.Curso;
 
 public class CursoMapper {
