+++ src/main/java/asist/io/service/IInscripcionService.java
@@ -1,7 +1,7 @@
 package asist.io.service;
 
-import asist.io.dto.InscripcionGetDTO;
-import asist.io.dto.InscripcionPostDTO;
+import asist.io.dto.inscripcionDTO.InscripcionGetDTO;
+import asist.io.dto.inscripcionDTO.InscripcionPostDTO;
 import asist.io.exception.ModelException;
 
 public interface IInscripcionService {
