+++ src/test/java/asist/io/EstudianteServiceTest.java
@@ -1,6 +1,11 @@
 package asist.io;
 
-import asist.io.dto.*;
+import asist.io.dto.cursoDTO.CursoGetDTO;
+import asist.io.dto.cursoDTO.CursoPostDTO;
+import asist.io.dto.estudianteDTO.EstudianteGetDTO;
+import asist.io.dto.estudianteDTO.EstudiantePostDTO;
+import asist.io.dto.inscripcionDTO.InscripcionGetDTO;
+import asist.io.dto.inscripcionDTO.InscripcionPostDTO;
 import asist.io.exception.ModelException;
 import asist.io.service.ICursoService;
 import asist.io.service.IEstudianteService;
